pyproject.toml
[tool.poetry]
name = "marketplace"
version = "0.1.0"
description = ""
authors = ["Momo Lo <toto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
attrs = "^22.2.0"
dataset = "^1.6.0"
email-validator = "^1.3.1"
phonenumbers = "^8.13.7"
pydantic = "^1.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
